Reto-1/requirements.txt
langchain>=0.1.0
langgraph
langchain-openai
streamlit
python-dotenv
pydantic
